Extract chat icon helpers
